requirements.txt
altair==5.1.2
annotated-types==0.6.0
anyio==4.1.0
appdirs==1.4.4
argcomplete @ file:///private/tmp/python-argcomplete-20231119-5188-lp06as/argcomplete-3.1.6
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi @ file:///private/tmp/python-certifi-20231117-4919-2f0o9m/certifi-2023.11.17
cffi @ file:///private/tmp/cffi-20231014-5567-1hyue8f/cffi-1.16.0
charset-normalizer @ file:///private/tmp/python-charset-normalizer-20231123-5225-x00973/charset-normalizer-3.3.2
click @ file:///private/tmp/python-click-20231113-5463-1enrr1r/click-8.1.7
contourpy==1.2.0
cryptography @ file:///private/tmp/python-cryptography-20231127-6248-fisajh/cryptography-41.0.7
cycler==0.12.1
defusedxml==0.7.1
distlib @ file:///private/tmp/python-distlib-20231124-5231-2n0vb3/distlib-0.3.7
distro==1.8.0
docutils @ file:///private/tmp/docutils-20231015-5454-w1xze4/docutils-0.20.1
entrypoints==0.4
Faker==20.1.0
favicon==0.7.0
filelock @ file:///private/tmp/python-filelock-20231112-5622-k62fq8/filelock-3.13.1/dist/filelock-3.13.1-py3-none-any.whl#sha256=57dbda9b35157b05fb3e58ee91448612eb674172fab98ee235ccb0b5bee19a1c
fonttools==4.45.1
gguf==0.5.1
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.2
httpx==0.25.2
idna @ file:///private/tmp/python-idna-20231125-5279-xtqzsf/idna-3.6
importlib-metadata==6.8.0
Jinja2 @ file:///private/tmp/python-jinja-20231112-5434-1yb7qop/Jinja2-3.1.2
jira==3.5.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe @ file:///private/tmp/python-markupsafe-20231113-5538-1gridwx/MarkupSafe-2.1.3
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==10.1.0
numpy==1.24.4
oauthlib==3.2.2
openai==1.3.8
packaging @ file:///private/tmp/python-packaging-20231018-5368-zv724j/packaging-23.2
pandas==2.1.3
Pillow==10.1.0
platformdirs @ file:///private/tmp/python-platformdirs-20231124-8775-x8oz64/platformdirs-4.0.0
protobuf==4.25.1
psutil @ file:///private/tmp/python-psutil-20231015-5378-1ng7p6p/psutil-5.9.6
pyarrow==14.0.1
pycparser @ file:///private/tmp/pycparser-20231013-5787-1anpge2/pycparser-2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pyee==8.2.2
Pygments @ file:///private/tmp/pygments-20231121-5274-1rlyumb/pygments-2.17.2
pymdown-extensions==10.5
pyparsing==3.1.1
pyppeteer==1.0.2
python-dateutil @ file:///private/tmp/python-dateutil-20231119-5156-1u9krby/python-dateutil-2.8.2
python-decouple==3.8
pytz==2023.3.post1
pywatchman==1.4.1
PyYAML @ file:///private/tmp/pyyaml-20231018-5610-1rdh755/PyYAML-6.0.1
referencing==0.31.0
requests @ file:///private/tmp/python-requests-20231124-5682-96f89l/requests-2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.13.1
sentencepiece==0.1.98
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
st-annotated-text==4.0.1
streamlit==1.28.2
streamlit-aggrid==0.3.4.post3
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.5
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-option-menu==0.3.6
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tabulate @ file:///private/tmp/python-tabulate-20231113-17262-1dzbtg2/tabulate-0.9.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
utils==1.0.1
validators==0.22.0
watchdog==3.0.0
websockets==10.4
zipp==3.17.0
